requirements.txt
flask
python-whois
requests
beautifulsoup4
dnspython
textdistance
gunicorn